Cleanup EsyfovarselKafkaConfig.kt. Delete unused EsyfovarselConsumer.kt. Delete empty java folder
