+++ src/main/kotlin/no/nav/syfo/config/kafka/KafkaAivenConfig.kt
@@ -1,9 +1,12 @@
 package no.nav.syfo.config.kafka
 
+import no.nav.syfo.varsel.esyfovarsel.JacksonKafkaSerializer
 import org.apache.kafka.clients.CommonClientConfigs
 import org.apache.kafka.clients.consumer.ConsumerConfig
+import org.apache.kafka.clients.producer.ProducerConfig
 import org.apache.kafka.common.config.SaslConfigs
 import org.apache.kafka.common.config.SslConfigs
+import org.apache.kafka.common.serialization.StringSerializer
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Configuration
 import org.springframework.kafka.annotation.EnableKafka
@@ -89,6 +92,24 @@ class KafkaAivenConfig(
             put(
                 ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG,
                 "" + (10 * 1024 * 1024)
+            )
+        }
+    }
+
+    fun commonKafkaAivenProducerConfig(): HashMap<String, Any> {
+        return HashMap<String, Any>().apply {
+            putAll(commonKafkaAivenConfig())
+            put(
+                ProducerConfig.ACKS_CONFIG,
+                "all"
+            )
+            put(
+                ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG,
+                StringSerializer::class.java
+            )
+            put(
+                ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG,
+                JacksonKafkaSerializer::class.java
             )
         }
     }
